T20: agent-facing CLI with run and triage verbs
run: filtered specimen selection (prefix/family/tag), one environment per
invocation, child processes per backend x batch through ChildProcessLauncher,
merge via ResultMerger, comparison via ClusterDiffer, stable stdout JSON,
exit 0 only when nothing differs and nothing failed.

triage: ranks DIFFERENT comparisons grouped by widget family and probable
defect class, optimising for fewest investigations per defect fixed;
deterministic for identical and reordered input.

Selftest grows 53 to 57 checks; docs/visual-oracle/CLI.md documents every
verb, flag, exit code and the first real finding: the skija-proto fork
renders the whole Link family on an off-theme surface (99.1-99.5% of pixels
differ from native, one shared defect).

+++ tools/oracle/harness/src/org/eclipse/swt/visualoracle/impl/ResultMerger.java
@@ -74,7 +74,10 @@ public final class ResultMerger {
 		captures.sort(CAPTURE_ORDER);
 		comparisons.sort(COMPARISON_ORDER);
 		Map<String, Object> merged = new LinkedHashMap<>();
-		merged.put("schemaVersion", Integer.valueOf(ResultSchemaValidator.SUPPORTED_SCHEMA_VERSION));
+		// Long, not Integer: the strict validator accepts exactly what JsonParser
+		// produces for JSON integers, so an in-memory consumer of this map
+		// (T20 run verb) validates the same document that a file round-trip yields.
+		merged.put("schemaVersion", Long.valueOf(ResultSchemaValidator.SUPPORTED_SCHEMA_VERSION));
 		merged.put("generator", generator);
 		merged.put("environment", environment);
 		merged.put("captures", captures);
